+++ addon-test-support/-private/steps/power-select.js
@@ -1,11 +1,15 @@
 import { expect } from 'chai';
 import { assert }  from '@ember/debug';
-import { click, currentURL, pauseTest, settled } from '@ember/test-helpers';
-import { getIndexZero, isVisible, pause } from 'ember-cli-yadda-opinionated/test-support/-private/helpers';
+import { click } from '@ember/test-helpers';
+import { getIndexZero } from 'ember-cli-yadda-opinionated/test-support/-private/helpers';
 
 import {
   powerSelectFindTrigger,
   powerSelectFindOptions,
+  powerSelectFindSelectedOptions,
+  powerSelectFilterSelectedOptionsByText,
+  powerSelectRemoveSelectedOption,
+  powerSelectIsSelectedOptionDisabled,
   powerSelectIsDropdownExpanded,
   powerSelectExpand,
   powerSelectCollapse,
@@ -40,6 +44,22 @@ const steps = {
     }
   },
 
+  async "Then there should be (NO|no )?(?:(\\d+) )selected items? in the dropdown $opinionatedElement"(no, countStr, [collection/* , label, selector */]) {
+    assert("Don't use NO and number at the same time", !(no && countStr));
+    assert(`Expected a single element, but ${collection.length} found`, collection.length === 1);
+
+    const count =
+      no       ? 0                      :
+      countStr ? parseInt(countStr, 10) : 0;
+
+    assert(`Must use either "NO" or number`, count !== undefined);
+
+    const trigger = powerSelectFindTrigger(collection[0]);
+    const selectedOptions = powerSelectFindSelectedOptions(trigger);
+
+    expect(selectedOptions).to.have.length(count);
+  },
+
   async "Then (?:(?:a|an|the) )?(?:(\\d+)(?:st|nd|rd|th) |(first|second|third|fourth|fifth|sixth|seventh|eighth|ninth|tenth) )?item in the dropdown $opinionatedElement should (NOT |not )?(?:have text|say) \"(.*)\""(indexOneStr, ordinal, [collection], not, text) {
     assert(`Expected a single element, but ${collection.length} found.`, collection.length === 1);
 
@@ -60,6 +80,47 @@ const steps = {
       : expect(options[indexZero].textContent.trim()).equal(text);
 
     // Restore the original state
+    if (!isExpanded) {
+      await powerSelectCollapse(trigger);
+    }
+  },
+
+  async "Then (?:(?:a|an|the) )?(?:(\\d+)(?:st|nd|rd|th) |(first|second|third|fourth|fifth|sixth|seventh|eighth|ninth|tenth) )?item in the dropdown $opinionatedElement should be disabled"(indexOneStr, ordinal, [collection]) {
+    assert(`Expected a single element, but ${collection.length} found.`, collection.length === 1);
+
+    const trigger = powerSelectFindTrigger(collection[0]);
+    const isExpanded = powerSelectIsDropdownExpanded(trigger);
+    const indexZero = getIndexZero(ordinal, indexOneStr, 0);
+
+    if (!isExpanded) {
+      await powerSelectExpand(trigger);
+    }
+
+    const options = powerSelectFindOptions(trigger);
+
+    assert(`Expected the dropdown to have at least ${indexZero + 1} elements.`, options.length >= indexZero + 1);
+
+    expect(options[indexZero]).to.have.attr('aria-disabled');
+
+    if (!isExpanded) {
+      await powerSelectCollapse(trigger);
+    }
+  },
+
+  async "Then (?:(?:a|an|the) )?(?:(\\d+)(?:st|nd|rd|th) |(first|second|third|fourth|fifth|sixth|seventh|eighth|ninth|tenth) )?item \"(.+)\" in the dropdown $opinionatedElement should be disabeld"(indexOneStr, ordinal, text, [collection]) {
+    assert(`Expected a single element, but ${collection.length} found.`, collection.length === 1);
+
+    const trigger = powerSelectFindTrigger(collection[0]);
+    const isExpanded = powerSelectIsDropdownExpanded(trigger);
+    const indexZero = getIndexZero(ordinal, indexOneStr, 0);
+
+    if (!isExpanded) {
+      await powerSelectExpand(trigger);
+    }
+
+    const option = powerSelectFindOptionByValueOrSelector(trigger, text, indexZero);
+    expect(option).to.have.attr('aria-disabled');
+
     if (!isExpanded) {
       await powerSelectCollapse(trigger);
     }
@@ -98,10 +159,68 @@ const steps = {
     return click(option);
   },
 
+  async "When I deselect (?:(?:a|an|the) )?(?:(\\d+)(?:st|nd|rd|th) |(first|second|third|fourth|fifth|sixth|seventh|eighth|ninth|tenth) )?selected item in the dropdown $opinionatedElement"(indexOneStr, ordinal, [collection]) {
+    assert(`Expected a single element, but ${collection.length} found.`, collection.length === 1);
+
+    const trigger = powerSelectFindTrigger(collection[0]);
+    const indexZero = getIndexZero(ordinal, indexOneStr, 0);
+    const selectedOptions = powerSelectFindSelectedOptions(trigger);
+
+    assert(`Expected at least ${indexZero + 1} selected option, but ${selectedOptions.length} found`, selectedOptions.length >= indexZero + 1);
+
+    return powerSelectRemoveSelectedOption(selectedOptions[indexZero]);
+  },
+
+  async "When I deselect (?:(?:a|an|the) )?(?:(\\d+)(?:st|nd|rd|th) |(first|second|third|fourth|fifth|sixth|seventh|eighth|ninth|tenth) )?selected item \"(.+)\" in the dropdown $opinionatedElement"(indexOneStr, ordinal, text, [collection]) {
+    assert(`Expected a single element, but ${collection.length} found.`, collection.length === 1);
+
+    const trigger = powerSelectFindTrigger(collection[0]);
+    const indexZero = getIndexZero(ordinal, indexOneStr, 0);
+
+    const selectedOptions = powerSelectFindSelectedOptions(trigger);
+    const filteredSelectedOptions = powerSelectFilterSelectedOptionsByText(selectedOptions, text);
+
+    assert(`Expected at least ${indexZero + 1} selected option to have text "${text}", but ${filteredSelectedOptions.length} found`, filteredSelectedOptions.length >= indexZero + 1);
+
+    return powerSelectRemoveSelectedOption(filteredSelectedOptions[indexZero]);
+  },
+
+  async "Then (?:(?:a|an|the) )?(?:(\\d+)(?:st|nd|rd|th) |(first|second|third|fourth|fifth|sixth|seventh|eighth|ninth|tenth) )?selected item in the dropdown $opinionatedElement should (NOT |not )?be disabled"(indexOneStr, ordinal, [collection], not) {
+    assert(`Expected a single element, but ${collection.length} found.`, collection.length === 1);
+
+    const trigger = powerSelectFindTrigger(collection[0]);
+    const indexZero = getIndexZero(ordinal, indexOneStr, 0);
+    const selectedOptions = powerSelectFindSelectedOptions(trigger);
+
+    assert(`Expected at least ${indexZero + 1} selected option, but ${selectedOptions.length} found`, selectedOptions.length >= indexZero + 1);
+
+    const result = powerSelectIsSelectedOptionDisabled(selectedOptions[indexZero]);
+
+    return not ? !result : result;
+  },
+
+  async "Then (?:(?:a|an|the) )?(?:(\\d+)(?:st|nd|rd|th) |(first|second|third|fourth|fifth|sixth|seventh|eighth|ninth|tenth) )?selected item \"(.+)\" in the dropdown $opinionatedElement should (NOT |not )?be disabled"(indexOneStr, ordinal, text, [collection], not) {
+    assert(`Expected a single element, but ${collection.length} found.`, collection.length === 1);
+
+    const trigger = powerSelectFindTrigger(collection[0]);
+    const indexZero = getIndexZero(ordinal, indexOneStr, 0);
+
+    const selectedOptions = powerSelectFindSelectedOptions(trigger);
+    const filteredSelectedOptions = powerSelectFilterSelectedOptionsByText(selectedOptions, text);
+
+    assert(`Expected at least ${indexZero + 1} selected option to have text "${text}", but ${filteredSelectedOptions.length} found`, filteredSelectedOptions.length >= indexZero + 1);
+
+    const result = powerSelectIsSelectedOptionDisabled(filteredSelectedOptions[indexZero]);
+
+    return not ? !result : result;
+  },
+
   "Then the dropdown $opinionatedElement should have \"(.*)\" selected"([collection], text) {
     assert(`Expected a single element, but ${collection.length} found.`, collection.length === 1);
 
     const trigger = powerSelectFindTrigger(collection[0]);
+
+    assert('Power Select Trigger not found', !collection.length);
 
     expect(trigger).to.have.trimmed.text(text);
   },